+++ noodles-vcf/src/indexer.rs
@@ -27,7 +27,7 @@ where
     let mut record = RecordBuf::default();
     let mut start_position = reader.get_ref().virtual_position();
 
-    while reader.read_record(&header, &mut record)? != 0 {
+    while reader.read_record_buf(&header, &mut record)? != 0 {
         let end_position = reader.get_ref().virtual_position();
         let chunk = Chunk::new(start_position, end_position);